apps/job_analyzer/requirements.txt
google-generativeai
requests
beautifulsoup4
python-dotenv
flask
markdown
playwright
